+++ src/Physics/CollisionDectector.cpp
@@ -40,10 +40,10 @@ bool CollisionDectector::CircleVsCircle(Circle a, Circle b){
     // ABS(R0 - R1) <= SQRT((x0 - x1)^2 + (y0 - y1)^2) <= (R0 + R1)
     // Raised all sides by a power of 2 removes the SQRT operator
 
-    int lowerBound = pow((a.Radius - b.Radius),2);
-    int upperBound = pow((a.Radius + b.Radius),2);
+    float lowerBound = pow((a.Radius - b.Radius),2);
+    float upperBound = pow((a.Radius + b.Radius),2);
 
-    int distance = pow((a.Position.Coordinates.X - b.Position.Coordinates.X), 2) + pow((a.Position.Coordinates.Y - b.Position.Coordinates.Y), 2);
+    float distance = pow((a.Position.Coordinates.X - b.Position.Coordinates.X), 2) + pow((a.Position.Coordinates.Y - b.Position.Coordinates.Y), 2);
 
     return lowerBound <= distance && distance <= upperBound;
 }
